requirements.txt
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
coverage==7.9.1
decorator==5.2.1
defusedxml==0.7.1
dj-database-url==3.0.1
Django==5.2.3
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
docopt==0.6.2
drf-spectacular==0.28.0
drf-yasg==1.21.10
executing==2.2.0
fastjsonschema==2.21.1
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.3
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
packaging==25.0
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
platformdirs==4.3.8
pluggy==1.6.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.11
pure_eval==0.2.3
Pygments==2.19.1
pytest==8.4.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
pywin32==311
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.28.0
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
stack-data==0.6.3
tinycss2==1.4.0
tornado==6.5.1
traitlets==5.14.3
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
